+++ dotnet/tests/Microsoft.Agents.AI.Foundry.Hosting.UnitTests/FoundryToolboxTests.cs
@@ -9,13 +9,12 @@ using System.Text;
 using System.Threading.Tasks;
 using Azure.AI.Projects;
 using Azure.AI.Projects.Agents;
-using Microsoft.Agents.AI.Foundry.Hosting;
 using Microsoft.Extensions.AI;
 
 #pragma warning disable OPENAI001
 #pragma warning disable AAIP001
 
-namespace Microsoft.Agents.AI.Foundry.UnitTests;
+namespace Microsoft.Agents.AI.Foundry.Hosting.UnitTests;
 
 /// <summary>
 /// Unit tests for the <see cref="FoundryToolbox"/> class.
