+++ dotnet/tests/Microsoft.Agents.AI.Foundry.Hosting.UnitTests/TestData/ToolboxVersionResponse.json
