+++ dotnet/tests/Microsoft.Agents.AI.Foundry.Hosting.UnitTests/WorkflowTestAgents.cs
@@ -1,0 +1,96 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using System.Text.Json;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.AI;
+using MeaiTextContent = Microsoft.Extensions.AI.TextContent;
+
+namespace Microsoft.Agents.AI.Foundry.Hosting.UnitTests;
+
+/// <summary>
+/// A test agent that streams a single text update.
+/// </summary>
+internal sealed class StreamingTextAgent(string id, string responseText) : AIAgent
+{
+    public new string Id => id;
+
+    protected override async IAsyncEnumerable<AgentResponseUpdate> RunCoreStreamingAsync(
+        IEnumerable<ChatMessage> messages,
+        AgentSession? session,
+        AgentRunOptions? options,
+        [EnumeratorCancellation] CancellationToken cancellationToken = default)
+    {
+        yield return new AgentResponseUpdate
+        {
+            MessageId = $"msg_{id}",
+            Contents = [new MeaiTextContent(responseText)]
+        };
+
+        await Task.CompletedTask;
+    }
+
+    protected override Task<AgentResponse> RunCoreAsync(
+        IEnumerable<ChatMessage> messages,
+        AgentSession? session,
+        AgentRunOptions? options,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<AgentSession> CreateSessionCoreAsync(
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<JsonElement> SerializeSessionCoreAsync(
+        AgentSession session,
+        JsonSerializerOptions? jsonSerializerOptions,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<AgentSession> DeserializeSessionCoreAsync(
+        JsonElement serializedState,
+        JsonSerializerOptions? jsonSerializerOptions,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+}
+
+/// <summary>
+/// A test agent that always throws an exception during streaming.
+/// </summary>
+internal sealed class ThrowingStreamingAgent(string id, Exception exception) : AIAgent
+{
+    public new string Id => id;
+
+    protected override IAsyncEnumerable<AgentResponseUpdate> RunCoreStreamingAsync(
+        IEnumerable<ChatMessage> messages,
+        AgentSession? session,
+        AgentRunOptions? options,
+        CancellationToken cancellationToken = default) =>
+        throw exception;
+
+    protected override Task<AgentResponse> RunCoreAsync(
+        IEnumerable<ChatMessage> messages,
+        AgentSession? session,
+        AgentRunOptions? options,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<AgentSession> CreateSessionCoreAsync(
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<JsonElement> SerializeSessionCoreAsync(
+        AgentSession session,
+        JsonSerializerOptions? jsonSerializerOptions,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+
+    protected override ValueTask<AgentSession> DeserializeSessionCoreAsync(
+        JsonElement serializedState,
+        JsonSerializerOptions? jsonSerializerOptions,
+        CancellationToken cancellationToken = default) =>
+        throw new NotImplementedException();
+}
